feat: Serve both gRPC and HTTP server
